#8564-TestCheckUsersCanSeeBatchEditButton : created test and refactored classes
