+++ sormas-e2e-tests/src/main/java/org/sormas/e2etests/pages/application/events/EventParticipantsPage.java
@@ -52,29 +52,16 @@ public class EventParticipantsPage {
   public static final By SELECT_PERSON_SEARCH_BUTTON_AT_ADD_PARTICIPANT = By.id("actionSearch");
   public static final By SELECT_FIRST_PERSON_IN_SEARCHED_LIST_FROM_ADD_PARTICIPANT =
       By.cssSelector("[scroll] [aria-live] .v-grid-body tr:nth-of-type(1)");
-  public static final By PERSON_ID_IN_EVENT_PARTICIPANTS_COLUMN_HEADER =
-      By.cssSelector(
-          "th:nth-of-type(2) > .v-grid-column-default-header-content.v-grid-column-header-content");
-  public static final By CASE_GRID_RESULTS_ROWS = By.cssSelector("[role=rowgroup] tr a");
-  public static final By PERSON_ID_IN_EVENT_PARTICIPANTS_RESULTS_ROWS =
-      By.cssSelector("tr[role='row'] > td:nth-of-type(2) a");
   public static final By CONFIRM_BUTTON_FOR_SELECT_PERSON_FROM_ADD_PARTICIPANTS_WINDOW =
       By.cssSelector("[scroll] [role='dialog']:nth-of-type(5) #commit");
   public static final By DELETE_EVENT_PARTICIPANT_BUTTTON = By.cssSelector("div#deleteUndelete");
-  public static final By CONFIRM_DELETION_OF_EVENT_PARTICIPANT = By.id("actionConfirm");
   public static final By CONFIRM_ACTION = By.id("actionConfirm");
-  public static final By POPUP_DISCARD_CHANGES_BUTTON = By.id("actionCancel");
   public static final By POPUP_CANCEL_ACTION_BUTTON = By.id("unsavedChanges.cancel");
   public static final By EVENT_PARTICIPANT_UUID =
       By.cssSelector("tr:nth-of-type(1) > td:nth-of-type(1) >a");
-  public static final By DESCRIPTION_OF_CREATE_CASE_FROM_EVENT_PARTICIPANT =
-      By.xpath(
-          "//div[@class='v-window-outerheader'][contains(.,'Create case from event participant with positive test result?')]");
   public static final By ARCHIVE_EVENT_PARTICIPANT_BUTTON = By.id("archiveDearchive");
   public static final By EVENT_PARTICIPANT_DISPLAY_FILTER_COMBOBOX =
       By.cssSelector("[id='relevanceStatusFilter'] [class='v-filterselect-button']");
-  public static final By EVENT_PARTICIPANT_DISPLAY_FILTER_INPUT =
-      By.cssSelector("[id='relevanceStatusFilter'] input");
   public static final By EVENT_TAB = By.id("tab-events-data");
   public static final By DEARCHIVE_REASON_TEXT_AREA = By.cssSelector(".popupContent textarea");
   public static final By CONFIRM_DEARCHIVE_BUTTON = By.id("actionConfirm");
@@ -82,6 +69,9 @@ public class EventParticipantsPage {
   public static final By PASSPORT_NUMBER_INPUT = By.id("passportNumber");
   public static final By INVOLVEMENT_DESCRIPTION_INPUT = By.id("involvementDescription");
   public static By NOTIFICATION_EVENT_PARTICIPANT = By.cssSelector(".v-Notification-description");
+  public static final By ENTER_BULK_EDIT_MODE_BUTTON = By.id("actionEnterBulkEditMode");
+  public static final By EVENT_PARTICIPANTS_GRID =
+      By.xpath("//div[@class='v-grid v-widget v-has-width v-has-height']");
 
   public static final By getEventParticipantByPersonUuid(String uuid) {
     return By.xpath(String.format("//a[@title='%s']//parent::td//parent::tr//td[1]/a", uuid));
